Ported darwin C-code to Objective-C, separating all three implementations into individual files
